Bring the command reference back in step, and number the ADRs
`commands.md` landed while this branch was in flight and documents, honestly,
four commands as **broken — cannot be run**. Three of those are `check`,
`policy list` and `config show`, which this branch fixed; so are
`profile remove --target` and `memory list --tag`. The callouts go, because a
reference that warns about a defect which no longer exists costs a reader the
same trip as one that hides a defect that does.

`status`, `deploy` and the new `sync targets` are documented as what they now
are: commands naming a target, with `--profile` narrowing.

The ADR directory needed sorting out, and not only by this branch. Upstream
had **three** files whose heading claimed ADR-040, two of them named `039-`,
and one of those in no index at all — parallel branches each taking the next
free number. Under the index, which is the authority: the cross-origin ADR is
039 and its heading is corrected; the orphaned owner ADR becomes 042; the two
written here become 043 and 044. Every reference in `src/` and `docs/` is
repointed by meaning rather than by number.

`docs.test.ts` now checks all three places a number lives — filename, heading,
and the row in `README.md` — because nothing did, which is how three files
came to claim one number. A reference like "(ADR-040)" in a comment is not
merely stale when that happens, it is ambiguous.

`selection.ts` went one line over the budget with `knowledge` and `sync` both
landing in it. `nearest` is the piece with no relationship to the rest: it is
string distance, it has no idea what a flag is, and it is the only part of
that file that does not grow when a command is added.
