+++ src/cli/selection.ts
@@ -10,6 +10,7 @@ import {
   targetsByName,
 } from '#profile';
 import type { Flags } from './argv.ts';
+import { nearest } from './nearest.ts';
 
 /**
  * Which commands must name a profile and a target, and which flags each accepts.
@@ -361,40 +362,4 @@ export function assertKnownFlags(first: string, second: string | undefined, flag
         `\n  Accepts: ${[...allowed].sort().map((name) => `--${name}`).join(' ')}`,
     );
   }
-}
-
-/**
- * The closest accepted flag, when there is an obviously close one.
- *
- * One edit away, or one transposition — enough for `--porfile` and `--taget`,
- * and short of guessing at something the operator did not mean. A wrong guess
- * here costs more than no guess: it sends them to a flag that is not the answer.
- */
-function nearest(given: string, allowed: ReadonlySet<string>): string | undefined {
-  for (const candidate of allowed) {
-    if (Math.abs(candidate.length - given.length) > 1) continue;
-    if (distance(given, candidate) <= 1) return candidate;
-    if (sorted(given) === sorted(candidate)) return candidate;
-  }
-  return undefined;
-}
-
-const sorted = (text: string): string => [...text].sort().join('');
-
-function distance(a: string, b: string): number {
-  let row = Array.from({ length: b.length + 1 }, (_, index) => index);
-
-  for (let i = 1; i <= a.length; i++) {
-    const next = [i];
-    for (let j = 1; j <= b.length; j++) {
-      next[j] = Math.min(
-        row[j]! + 1,
-        next[j - 1]! + 1,
-        row[j - 1]! + (a[i - 1] === b[j - 1] ? 0 : 1),
-      );
-    }
-    row = next;
-  }
-
-  return row[b.length]!;
 }
